Add MongoDB database setup and CRUD functionality
This commit adds the necessary files and configurations to set up a MongoDB database and implement basic CRUD functionality for articles. It also includes a README file with instructions on how to set up and use the database.

Additionally, this commit welcomes a new contributor to the repository, Gerardo Ormeno. In the next commit, new tasks will be provided for him to work on.

Fixes #123